fixed preemption for will run calls

+++ src/plugins/select/bluegene/plugin/bg_job_place.h
@@ -61,7 +61,7 @@
  */
 extern int submit_job(struct job_record *job_ptr, bitstr_t *bitmap,
 		      uint32_t min_nodes, uint32_t max_nodes,
-		      uint32_t req_nodes, int mode,
+		      uint32_t req_nodes, uint16_t mode,
 		      List preemptee_candidates,
 		      List *preemptee_job_list);
